add download file

+++ SpeakTogether/Service/FileStorage/LocalMaterialStorageService.cs
@@ -6,34 +6,44 @@ namespace SpeakTogether.Service.FileStorage
 {
     public class LocalMaterialStorageService : IFileStorage
     {
-        private readonly string _basePath = Path.Combine(Directory.GetCurrentDirectory(), "AppData/Materials");
-        private readonly ISpeakTogetherDbContext speakTogetherDbContext;
+        private readonly string _basePath = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot/materials");
 
-        public LocalMaterialStorageService(ISpeakTogetherDbContext speakTogetherDbContext)
+        public Task DeleteFileAsync(string fileName)
         {
-            this.speakTogetherDbContext = speakTogetherDbContext;
-        }
+            var fullPath = Path.Combine(_basePath, fileName);
 
-        public Task DeleteFileAsync(string filePath)
-        {
-            var fullPath = Path.Combine(_basePath, filePath);
-            File.Delete(fullPath);
+            if (File.Exists(fullPath))
+                File.Delete(fullPath);
+
             return Task.CompletedTask;
         }
 
-        public Task<Stream> GetFileAsync(string filePath)
+        public Task<Stream?> GetFileAsync(string fileName)
         {
-            var fullPath = Path.Combine(_basePath, filePath);
-            return Task.FromResult<Stream>(File.OpenRead(fullPath));
+            var fullPath = Path.Combine(_basePath, fileName);
+
+            if (!File.Exists(fullPath))
+                return Task.FromResult<Stream?>(null);
+
+            Stream stream = File.OpenRead(fullPath);
+
+            return Task.FromResult<Stream?>(stream);
         }
 
         public async Task<string> SaveFileAsync(IFormFile file)
         {
-            var path = Path.Combine(_basePath, file.FileName);
-            using var stream = new FileStream(path, FileMode.Create);
+            if (!Directory.Exists(_basePath))
+                Directory.CreateDirectory(_basePath);
 
+            var fileName = Guid.NewGuid() + Path.GetExtension(file.FileName);
+
+            var fullPath = Path.Combine(_basePath, fileName);
+
+            using var stream = new FileStream(fullPath, FileMode.Create);
             await file.CopyToAsync(stream);
-            return path;
+
+            return fileName;
         }
+
     }
 }
